# tools/openapi.yaml
...
openapi: 3.0.3
info:
  title: VCell API
  description: VCell API
  termsOfService: Your terms of service here
  contact:
    name: VCell API Support
    url: https://vcell.org
    email: vcell_support@uchc.com
  license:
    name: MIT
    url: http://opensource.org/licenses/MIT
  version: 1.0.1
servers:
- url: https://vcell.cam.uchc.edu
paths:
...
  /api/v1/bioModel/save:
    post:
      tags:
      - Bio Model Resource
      summary: Save's the given BioModel. Optional parameters of name and simulations
        to update due to math changes. Returns saved BioModel as VCML.
      operationId: saveBioModel
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveBioModel'
      responses:
        "200":
          description: OK
          content:
            application/xml:
              schema:
                type: string
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "422":
          description: Unprocessable content submitted
        "555":
          description: Data Access Exception
      security:
      - openId:
        - user
  /api/v1/bioModel/{bioModelID}:
    get:
      tags:
      - Bio Model Resource
      summary: Get BioModel.
      operationId: getBioModel
      parameters:
      - name: bioModelID
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BioModel'
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "404":
          description: Not found
        "555":
          description: Data Access Exception
    delete:
      tags:
      - Bio Model Resource
      summary: Delete the BioModel from VCell's database.
      operationId: deleteBioModel
      parameters:
      - name: bioModelID
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "555":
          description: Data Access Exception
  /api/v1/bioModel/{bioModelID}/vcml_download:
    get:
      tags:
      - Bio Model Resource
      summary: Get the BioModel in VCML format.
      operationId: getBioModelVCML
      parameters:
      - name: bioModelID
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            text/xml:
              schema:
                type: string
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "404":
          description: Not found
        "555":
          description: Data Access Exception
...
  /api/v1/publications:
    get:
      tags:
      - Publication Resource
      summary: Get all publications
      operationId: getPublications
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Publication'
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
    put:
      tags:
      - Publication Resource
      summary: Update publication
      operationId: updatePublication
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Publication'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Publication'
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - openId:
        - user
    post:
      tags:
      - Publication Resource
      summary: Create publication
      operationId: createPublication
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Publication'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                format: int64
                type: integer
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - openId:
        - user
  /api/v1/publications/{id}:
    get:
      tags:
      - Publication Resource
      summary: Get publication by ID
      operationId: getPublicationById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Publication'
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
    delete:
      tags:
      - Publication Resource
      summary: Delete publication
      operationId: deletePublication
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "204":
          description: No Content
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "404":
          description: Not found
      security:
      - openId:
        - user
...
components:
  schemas:
...
    Application:
      type: object
...
    BioModel:
      type: object
      properties:
        bmKey:
          type: string
        name:
          type: string
        privacy:
          format: int32
          type: integer
        groupUsers:
          type: array
          items:
            type: string
        savedDate:
          format: int64
          type: integer
        annot:
          type: string
        branchID:
          type: string
        physModelKey:
          type: string
        ownerName:
          type: string
        ownerKey:
          type: string
        simulationKeyList:
          type: array
          items:
            $ref: '#/components/schemas/KeyValue'
        applications:
          type: array
          items:
            $ref: '#/components/schemas/Application'
    BiomodelRef:
      type: object
      properties:
        bmKey:
          format: int64
          type: integer
        name:
          type: string
        ownerName:
          type: string
        ownerKey:
          format: int64
          type: integer
        versionFlag:
          format: int32
          type: integer
...
    Date:
      format: date
      type: string
      example: 2022-03-10
...
    KeyValue:
      type: object
      properties:
        value:
          type: number
    MathmodelRef:
      type: object
      properties:
        mmKey:
          format: int64
          type: integer
        name:
          type: string
        ownerName:
          type: string
        ownerKey:
          format: int64
          type: integer
        versionFlag:
          format: int32
          type: integer
...
    Publication:
      type: object
      properties:
        pubKey:
          format: int64
          type: integer
        title:
          type: string
        authors:
          type: array
          items:
            type: string
        year:
          format: int32
          type: integer
        citation:
          type: string
        pubmedid:
          type: string
        doi:
          type: string
        endnoteid:
          format: int32
          type: integer
        url:
          type: string
        wittid:
          format: int32
          type: integer
        biomodelRefs:
          type: array
          items:
            $ref: '#/components/schemas/BiomodelRef'
        mathmodelRefs:
          type: array
          items:
            $ref: '#/components/schemas/MathmodelRef'
        date:
          $ref: '#/components/schemas/Date'
    SaveBioModel:
      required:
      - bioModelXML
      type: object
      properties:
        bioModelXML:
          pattern: \S
          type: string
        name:
          type: string
          nullable: true
        simsRequiringUpdates:
          type: array
          items:
            type: string
          nullable: true
...
  securitySchemes:
    openId:
      type: openIdConnect
      description: Authentication
      openIdConnectUrl: https://dev-dzhx7i2db3x3kkvq.us.auth0.com/.well-known/openid-configuration